requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.7.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Environment Variables
python-dotenv==1.0.0

# Production ASGI Server
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For TestClient
mongomock-motor==0.0.21  # Mock MongoDB for tests

# Code Quality
black==23.12.1          # Code formatting
flake8==7.0.0           # Linting
mypy==1.8.0             # Type checking
isort==5.13.2           # Import sorting

# Development Tools
ipython==8.19.0         # Enhanced Python shell
pre-commit==3.6.0       # Git hooks
